System.Random.TF

Description

This module exports System.Random.TF.Gen and System.Random.TF.Init modules without exporting the alternative RandomGen class from System.Random.TF.Gen. To use this class and the Random instances written for it, please import System.Random.TF.Gen and System.Random.TF.Instances directly.

Documentation

data TFGen

The generator type

seedTFGen :: (Word64, Word64, Word64, Word64) -> TFGen

Create a generator from a random seed.

newTFGen :: IO TFGen

Derive a new generator instance from the global RNG using split. This is the default way of obtaining a new RNG instance. Initial generator is seeded using mkSeedUnix on UNIX, and mkSeedTime otherwise. This should be eventually replaced with proper seeding.

mkTFGen :: Int -> TFGen

Quick and dirty way of creating a deterministically seeded generator.

mkSeedTime :: IO (Word64, Word64, Word64, Word64)

Use system time create the random seed. This method of seeding may not be relible.

mkSeedUnix :: IO (Word64, Word64, Word64, Word64)

Use the UNIX special file /dev/urandom to create the seed. Inspired by random-mwc.